+++ TinyBite/mocks/mockChatGroup.ts
@@ -1,92 +1,100 @@
-import { ChatMessage } from "@/types/chat";
+import { ChatMessageSchema } from "@/types/chat.types";
 
-export const mockChatGroup: ChatMessage[] = [
+export const mockChatGroup: ChatMessageSchema[] = [
   {
-    id: "12",
-    type: "image",
-    senderId: 4,
-    nickname: "츄비",
-    imageUrl: "https://picsum.photos/401/301",
+    messageId: 12,
+    messageType: "IMAGE",
     createdAt: "2025-11-22T18:06:50.000Z",
-  },
-  {
-    id: "11",
-    type: "image",
     senderId: 4,
     nickname: "츄비",
-    imageUrl: "https://picsum.photos/400/300",
+    isMine: false,
+    imageUrl: "https://picsum.photos/401/301",
+  },
+  {
+    messageId: 11,
+    messageType: "IMAGE",
     createdAt: "2025-11-22T18:06:40.000Z",
-  },
-  {
-    id: "10",
-    type: "text",
-    senderId: 1,
-    nickname: "나",
-    text: "입금 부탁드려요! 확인하는 대로 바로 구매 진행할게요. 입금 완료하시면 한 번만 톡 남겨주세요! 🙏",
-    createdAt: "2025-11-22T18:06:20.000Z",
-  },
-  {
-    id: "9",
-    type: "text",
-    senderId: 1,
-    nickname: "나",
-    text: "우리은행 1002-345-678912",
-    createdAt: "2025-11-22T18:06:10.000Z",
-  },
-  {
-    id: "8",
-    type: "text",
     senderId: 4,
     nickname: "츄비",
-    text: "안녕하세요! 계좌번호 주시면 입금할게요.",
-    createdAt: "2025-11-22T18:06:00.000Z",
+    isMine: false,
+    imageUrl: "https://picsum.photos/400/300",
   },
   {
-    id: "7",
-    type: "system",
-    systemMessage: "츄비님이 파티에 참가하셨습니다.",
-    createdAt: "2025-11-22T18:05:50.000Z",
-  },
-  {
-    id: "6",
-    type: "text",
+    messageId: 10,
+    messageType: "TEXT",
+    createdAt: "2025-11-22T18:06:20.000Z",
     senderId: 1,
     nickname: "나",
-    text: "한 분 더 들어오시면 바로 주문 진행하겠습니다!",
+    isMine: true,
+    text: "입금 부탁드려요! 확인하는 대로 바로 구매 진행할게요. 입금 완료하시면 한 번만 톡 남겨주세요! 🙏",
+  },
+  {
+    messageId: 9,
+    messageType: "TEXT",
+    createdAt: "2025-11-22T18:06:10.000Z",
+    senderId: 1,
+    nickname: "나",
+    isMine: true,
+    text: "우리은행 1002-345-678912",
+  },
+  {
+    messageId: 8,
+    messageType: "TEXT",
+    createdAt: "2025-11-22T18:06:00.000Z",
+    senderId: 4,
+    nickname: "츄비",
+    isMine: false,
+    text: "안녕하세요! 계좌번호 주시면 입금할게요.",
+  },
+  {
+    messageId: 7,
+    messageType: "SYSTEM",
+    createdAt: "2025-11-22T18:05:50.000Z",
+    systemMessage: "츄비님이 파티에 참가하셨습니다.",
+  },
+  {
+    messageId: 6,
+    messageType: "TEXT",
     createdAt: "2025-11-22T18:05:40.000Z",
+    senderId: 1,
+    nickname: "나",
+    isMine: true,
+    text: "한 분 더 들어오시면 바로 주문 진행하겠습니다!",
   },
   {
-    id: "5",
-    type: "text",
-    senderId: 3,
-    nickname: "엽떡조아",
-    text: "언제 주문하나요?",
+    messageId: 5,
+    messageType: "TEXT",
     createdAt: "2025-11-22T18:05:30.000Z",
-  },
-  {
-    id: "4",
-    type: "text",
     senderId: 3,
     nickname: "엽떡조아",
-    text: "안녕하세요",
+    isMine: false,
+    text: "언제 주문하나요?",
+  },
+  {
+    messageId: 4,
+    messageType: "TEXT",
     createdAt: "2025-11-22T18:05:20.000Z",
+    senderId: 3,
+    nickname: "엽떡조아",
+    isMine: false,
+    text: "안녕하세요",
   },
   {
-    id: "3",
-    type: "system",
-    systemMessage: "엽떡조아님이 파티에 참가하셨습니다.",
+    messageId: 3,
+    messageType: "SYSTEM",
     createdAt: "2025-11-22T18:05:10.000Z",
+    systemMessage: "엽떡조아님이 파티에 참가하셨습니다.",
   },
   {
-    id: "2",
-    type: "system",
-    systemMessage: "파티가 생성되었습니다.",
+    messageId: 2,
+    messageType: "SYSTEM",
     createdAt: "2025-11-22T18:05:05.000Z",
+    systemMessage: "파티가 생성되었습니다.",
   },
   {
-    id: "1",
-    type: "date",
-    systemMessage: "2025.11.22",
+    messageId: 1,
+    messageType: "DATE",
     createdAt: "2025-11-22T00:00:00.000Z",
+    date: "2025.11.22",
   },
 ];
